voice_agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit>=1.0.0
anthropic>=0.86.0
python-dotenv>=1.0.0
